+++ mex/international_projects/main.py
@@ -22,7 +22,7 @@ from mex.international_projects.transform import (
 from mex.pipeline import asset, run_job_in_process
 from mex.sinks import load
 from mex.wikidata.extract import (
-    get_organization_merged_id_by_query_with_transform_and_load,
+    get_merged_organization_id_by_query_with_transform_and_load,
 )
 
 
@@ -75,7 +75,7 @@ def international_projects_funding_sources_ids_by_query(
         international_projects_sources
     )
 
-    return get_organization_merged_id_by_query_with_transform_and_load(
+    return get_merged_organization_id_by_query_with_transform_and_load(
         wikidata_funding_sources_by_query, extracted_primary_source_wikidata
     )
 
@@ -92,7 +92,7 @@ def international_projects_partner_organization_ids_by_query(
         )
     )
 
-    return get_organization_merged_id_by_query_with_transform_and_load(
+    return get_merged_organization_id_by_query_with_transform_and_load(
         wikidata_partner_organizations_by_query, extracted_primary_source_wikidata
     )