+++ tests/wikidata/test_extract.py
@@ -1,7 +1,7 @@
 from mex.common.models import ExtractedPrimarySource
 from mex.common.wikidata.models.organization import WikidataOrganization
 from mex.wikidata.extract import (
-    get_organization_merged_id_by_query_with_transform_and_load,
+    get_merged_organization_id_by_query_with_transform_and_load,
 )
 
 
@@ -11,7 +11,7 @@ def test_get_organization_merged_id_by_query_with_transform_and_load(
 ) -> None:
     wikidata_primary_source = extracted_primary_sources["wikidata"]
 
-    returned = get_organization_merged_id_by_query_with_transform_and_load(
+    returned = get_merged_organization_id_by_query_with_transform_and_load(
         {"foo": wikidata_organization}, wikidata_primary_source
     )
     assert returned == {"foo": "ga6xh6pgMwgq7DC7r6Wjqg"}